Split out platform independent converter logic for BooleanToString value converter
